IGNITE-23807 Remove all DirectByteBufferStreamImpl except last (#11688)
