+++ modules/core/src/main/java/org/apache/ignite/internal/util/IgniteUtils.java
@@ -215,7 +215,6 @@ import org.apache.ignite.internal.compute.ComputeTaskCancelledCheckedException;
 import org.apache.ignite.internal.compute.ComputeTaskTimeoutCheckedException;
 import org.apache.ignite.internal.events.DiscoveryCustomEvent;
 import org.apache.ignite.internal.logger.IgniteLoggerEx;
-import org.apache.ignite.internal.managers.communication.GridIoManager;
 import org.apache.ignite.internal.managers.communication.GridIoPolicy;
 import org.apache.ignite.internal.managers.deployment.GridDeployment;
 import org.apache.ignite.internal.managers.deployment.GridDeploymentInfo;
@@ -10625,36 +10624,6 @@ public abstract class IgniteUtils {
 
             throw new IgniteInterruptedCheckedException(e);
         }
-    }
-
-    /**
-     * Defines which protocol version to use for
-     * communication with the provided node.
-     *
-     * @param ctx Context.
-     * @param nodeId Node ID.
-     * @return Protocol version.
-     * @throws IgniteCheckedException If node doesn't exist.
-     */
-    public static byte directProtocolVersion(GridKernalContext ctx, UUID nodeId) throws IgniteCheckedException {
-        assert nodeId != null;
-
-        ClusterNode node = ctx.discovery().node(nodeId);
-
-        if (node == null)
-            throw new IgniteCheckedException("Failed to define communication protocol version " +
-                "(has node left topology?): " + nodeId);
-
-        assert !node.isLocal();
-
-        Byte attr = node.attribute(GridIoManager.DIRECT_PROTO_VER_ATTR);
-
-        byte rmtProtoVer = attr != null ? attr : 1;
-
-        if (rmtProtoVer < GridIoManager.DIRECT_PROTO_VER)
-            return rmtProtoVer;
-        else
-            return GridIoManager.DIRECT_PROTO_VER;
     }
 
     /**
