+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/query/continuous/IgniteCacheContinuousQueryImmutableEntryTest.java
@@ -148,7 +148,7 @@ public class IgniteCacheContinuousQueryImmutableEntryTest extends GridCommonAbst
         e0.markFiltered();
 
         ByteBuffer buf = ByteBuffer.allocate(4096);
-        DirectMessageWriter writer = new DirectMessageWriter((byte)1);
+        DirectMessageWriter writer = new DirectMessageWriter();
 
         // Skip write class header.
         writer.onHeaderWritten();
@@ -157,7 +157,7 @@ public class IgniteCacheContinuousQueryImmutableEntryTest extends GridCommonAbst
         CacheContinuousQueryEntry e1 = new CacheContinuousQueryEntry();
         IgniteMessageFactoryImpl msgFactory =
                 new IgniteMessageFactoryImpl(new MessageFactory[]{new GridIoMessageFactory()});
-        e1.readFrom(ByteBuffer.wrap(buf.array()), new DirectMessageReader(msgFactory, (byte)1));
+        e1.readFrom(ByteBuffer.wrap(buf.array()), new DirectMessageReader(msgFactory));
 
         assertEquals(e0.cacheId(), e1.cacheId());
         assertEquals(e0.eventType(), e1.eventType());
